Make dependency on S3 optional, for easier self-hosting of admin (#147)
Right now, admin has a dependency on S3 in order to power custom logos.
This PR softens that dependency; S3 is only a dependency if you want to
use custom logos.

One known caveat in this implementation is that requesting an upload URL
marks the environment has having a logo defined. But the upload, which
the app immediately uses after requesting the URL, may fail. This isn't
a problem for this PR, because 1) we still have handling for GetObject
failing, and 2) for self-hosted instances that don't have s3 configured
correctly, the attempt to generate a presigned URL will fail beforehand.

Additionally, this PR clarifies/corrects some logs related to running
admin and looking for ADMIN_ env vars.