Add scimpatch and use SCIM data model as IR in user scim operations (#167)
This PR makes two changes to how SCIM user operations work:

1. Introduce scimpatch, which implements SCIM patch semantics, but with
exceptions to the spec made to accommodate Entra's noncompliant
behavior.
2. The SCIM data model now acts as an intermediate representation (IR).
SCIM operations operate on the SCIM data model representation of users,
which gets converted to/from the store layer's representation.

The main difficulty here is in accommodating, in a way that makes users
get details right by default, two big ways that Entra does not comply
with the SCIM specification:

1. The `active` property is supposed to be a boolean, but Entra sets it
as the JSON string `"True"` or `"False"` by default.
2. Entra makes up, in some but not all cases, its own syntax for paths
to nested fields. In particular, it sends requests with this kind of
path:


```
urn:ietf:params:scim:schemas:extension:enterprise:2.0:User:manager
```

With the intention being that this manipulates
`urn:ietf:params:scim:schemas:extension:enterprise:2.0:User` >
`manager`. This is illegal on two levels: the `.` in `2.0` should be a
field delimiter, and the `:` in `User:manager` is not a legal field
delimiter. There's no good logic here, either -- `:` can't be considered
a delimiter in general, because
`urn:ietf:params:scim:schemas:extension:enterprise:2.0:User` is meant to
be considered a single field.

There's no way to make Entra's logic make sense. Instead, this PR opts
to simply special-case this syntax.