auth: return 400 on repeated IDP-initiated SAML flows (#216)
This PR has auth return a 400 on replayed/repeated IDP-initated SAML
flows. We already have this logic in place for SP-initiated flows, where
we can catch it earlier in the flow.

When an assertion turns out to violate the `(saml_connection_id,
assertion_id)` uniqueness constraint, then store now returns a specific
error for that case.