Populate diagnostic RequestID and AssertionID even on invalid assertions (#219)
This PR improves error reporting on assertions that have an invalid
signature. Today, these assertions are incorrectly considered as having
been previously processed because they lack an assertion ID. In fact,
they do have such an ID, but we don't report it when the assertion has
an invalid signature.

With this PR, the SAML assertion's request and assertion IDs are now
populated on a preliminary basis prior to validating the signature. The
final data continues to be populated from the canonicalized data in the
happy path, but the diagnostic data for the unhappy path is now richer.