internal/saml: refactor Validate to return only result, error (#220)
This PR refactors internal/saml.Validate to return only a result, error
pair.

Right now, Validate returns a result, a set of "problems", and an error.
The problems indicate that the assertion is invalid, but in a way that
we can produce diagnostics about. With this PR, the diagnostic
information is always contained within a saml.ValidateError as the
returned error.

As part of this change, we now have something close to table-driven
tests for the diagnostic error cases.