internal/saml: refactor Validate to return only result, error (#223)
This PR is an amended version of #220, which was reverted in #222. The
amendment includes logic to populate RequestID, AssertionID, and
Assertion in the happy cases. Previous testing failed to detect that
happy-path saml flows did not have these fields populated, leading auth
to incorrectly conclude that assertions were being replayed.